+++ ui/src/loudml/components/TrainTimeJobButton.js
@@ -3,12 +3,13 @@ import {PropTypes} from 'prop-types'
 
 import CustomTimeJobButton from 'src/loudml/components/CustomTimeJobButton'
 
-const TrainTimeJobButton = ({
+export const TrainTimeJobButton = ({
     startLabel,
     stopLabel,
     onStart,
     onStop,
     running,
+    timeRange,
 }) => {
     
     return (
@@ -18,16 +19,23 @@ const TrainTimeJobButton = ({
             onStart={onStart}
             onStop={onStop}
             running={running}
+            selected={timeRange}
         />
     );
   }
 
+const {string, func, bool, shape} = PropTypes
+
 TrainTimeJobButton.propTypes = {
-    startLabel: PropTypes.string,
-    stopLabel: PropTypes.string,
-    onStart: PropTypes.func,
-    onStop: PropTypes.func,
-    running: PropTypes.bool,
+    startLabel: string,
+    stopLabel: string,
+    onStart: func,
+    onStop: func,
+    running: bool,
+    timeRange: shape({
+        lower: string.isRequired,
+        upper: string,
+        }),
 }
 
 export default TrainTimeJobButton