+++ ui/src/loudml/constants/index.js
@@ -30,6 +30,8 @@ export const DEFAULT_FEATURE = {
     io: 'io',
     anomaly_type: 'low_high',
     match_all: [],
+    scores: 'min_max',
+    transform: null,
 }
 
 export const DEFAULT_METRICS = [
@@ -44,9 +46,19 @@ export const DEFAULT_METRICS = [
 export const DEFAULT_IO  = [
     { text: 'in', value: 'i', },
     { text: 'out', value: 'o', },
-    { text: 'in/out', value: 'io', }
+    { text: 'in/out', value: 'io', },
 ]
                                     
+export const DEFAULT_SCORES  = [
+    { text: 'min/max', value: 'min_max', },
+    { text: 'normalize', value: 'normalize', },
+    { text: 'standardize', value: 'standardize', },
+]
+
+export const DEFAULT_TRANSFORM  = [
+    { text: 'diff', value: 'diff', },
+    { text: 'none', value: null, },
+]
 
 export const MODEL_CREATED = 'MODEL_CREATED';
 export const MODEL_UPDATED = 'MODEL_UPDATED';